refactor: `morphing` モジュールを作成
