+++ voicevox_engine/app/routers/morphing.py
@@ -13,16 +13,16 @@ from voicevox_engine.core.core_initializer import CoreManager
 from voicevox_engine.metas.Metas import StyleId
 from voicevox_engine.metas.MetasStore import MetasStore
 from voicevox_engine.model import AudioQuery
-from voicevox_engine.morphing import (
+from voicevox_engine.morphing.morphing import (
     MorphableTargetInfo,
     StyleIdNotFoundError,
     get_morphable_targets,
     is_morphable,
 )
-from voicevox_engine.morphing import (
+from voicevox_engine.morphing.morphing import (
     synthesis_morphing_parameter as _synthesis_morphing_parameter,
 )
-from voicevox_engine.morphing import synthesize_morphed_wave
+from voicevox_engine.morphing.morphing import synthesize_morphed_wave
 from voicevox_engine.tts_pipeline.tts_engine import TTSEngineManager
 from voicevox_engine.utility.path_utility import delete_file
